+++ tests/linux/unit/internals/support/internals_common.h
@@ -20,6 +20,7 @@
 #include "edhoc_transcript_hash_internal.h"
 #include "edhoc_mac_internal.h"
 #include "edhoc_plaintext_internal.h"
+#include "edhoc_cipher_internal.h"
 
 /* Defines ----------------------------------------------------------------- */
 /* Types and type definitions ---------------------------------------------- */
@@ -60,8 +61,6 @@ extern int comp_salt_3e2m(const struct edhoc_context *ctx, uint8_t *salt,
 			  size_t salt_len);
 extern int comp_encapsulate(struct edhoc_context *ctx);
 extern int comp_decapsulate(struct edhoc_context *ctx);
-extern int comp_keystream(const struct edhoc_context *ctx, uint8_t *keystream,
-			  size_t keystream_len);
 extern int comp_th_3(struct edhoc_context *ctx,
 		     const struct mac_context *mac_ctx, const uint8_t *ptxt,
 		     size_t ptxt_len);
@@ -89,8 +88,6 @@ extern int comp_prk_4e3m(struct edhoc_context *ctx, const void *private_key_id,
 			 size_t peer_public_key_length);
 extern int comp_salt_4e3m(const struct edhoc_context *ctx, uint8_t *salt,
 			  size_t salt_len);
-extern int comp_key_iv_aad_3(struct edhoc_context *ctx, uint8_t *iv,
-			     size_t iv_len, uint8_t *aad, size_t aad_len);
 extern int comp_th_4(struct edhoc_context *ctx,
 		     const struct mac_context *mac_ctx, const uint8_t *ptxt,
 		     size_t ptxt_len);
@@ -104,22 +101,14 @@ extern int prepare_plaintext_3(const struct mac_context *mac_ctx,
 			       const uint8_t *sign, size_t sign_len,
 			       uint8_t *ptxt, size_t ptxt_size,
 			       size_t *ptxt_len);
-extern int comp_aad_3_len(const struct edhoc_context *ctx, size_t *aad_3_len);
 extern int gen_msg_3(const uint8_t *ctxt, size_t ctxt_len, uint8_t *msg_3,
 		     size_t msg_3_size, size_t *msg_3_len);
 extern int parse_msg_3(const uint8_t *msg_3, size_t msg_3_len,
 		       const uint8_t **ctxt_3, size_t *ctxt_3_len);
-extern int decrypt_ciphertext_3(const struct edhoc_context *ctx,
-				const uint8_t *iv, size_t iv_len,
-				const uint8_t *aad, size_t aad_len,
-				const uint8_t *ctxt, size_t ctxt_len,
-				uint8_t *ptxt, size_t ptxt_len);
 extern int parse_plaintext_3(struct edhoc_context *ctx, const uint8_t *ptxt,
 			     size_t ptxt_len, struct plaintext *parsed_ptxt);
 extern int compute_plaintext_4_len(const struct edhoc_context *ctx,
 				   size_t *ptxt_4_len);
-extern int compute_key_iv_aad_4(struct edhoc_context *ctx, uint8_t *iv,
-				size_t iv_len, uint8_t *aad, size_t aad_len);
 extern int prepare_plaintext_4(const struct edhoc_context *ctx, uint8_t *ptxt,
 			       size_t ptxt_size, size_t *ptxt_len);
 extern int gen_msg_4(const uint8_t *ctxt, size_t ctxt_len, uint8_t *msg_4,