+++ tests/linux/unit/internals/test_internals_message4.c
@@ -172,8 +172,8 @@ TEST(internals_message4, compute_key_iv_aad_4_null)
 	uint8_t iv[13] = { 0 };
 	uint8_t aad[256] = { 0 };
 
-	int ret = compute_key_iv_aad_4(NULL, iv, ARRAY_SIZE(iv), aad,
-				       ARRAY_SIZE(aad));
+	int ret = edhoc_cipher_derive(NULL, iv, ARRAY_SIZE(iv), aad,
+				      ARRAY_SIZE(aad));
 	TEST_ASSERT_EQUAL(EDHOC_ERROR_INVALID_ARGUMENT, ret);
 }
 
@@ -188,8 +188,8 @@ TEST(internals_message4, compute_key_iv_aad_4_bad_state)
 	uint8_t iv[13] = { 0 };
 	uint8_t aad[256] = { 0 };
 
-	int ret = compute_key_iv_aad_4(&ctx, iv, ARRAY_SIZE(iv), aad,
-				       ARRAY_SIZE(aad));
+	int ret = edhoc_cipher_derive(&ctx, iv, ARRAY_SIZE(iv), aad,
+				      ARRAY_SIZE(aad));
 	TEST_ASSERT_EQUAL(EDHOC_ERROR_BAD_STATE, ret);
 
 	ret = edhoc_context_deinit(&ctx);